requirements.txt
numpy
pandas
scikit-learn
cupy-cuda12x
requests
pyarrow
